Remove unreachable code on `NimbleOptions`
I just started checking this library, and it looks great, thanks for your hard
work.

I _think_ these two cases of the private function `reduce_options/2` are not
reachable, sorry in advance if that's not the case. I did review the unit
tests, and there I don't see any code that could be reaching these functions.