Support a nesting level in NimbleOptions.docs/2 (#57)
The nesting level allows to nest all options in the generated
Markdown by the given level. This allows to have, for example,
a nested "nested_schema" schema and use NimbleOptions.docs/2
inside the documentation for another option:

    main_schema = [
      my_key: [
        type: {:or, [:string, keyword_list: nested_schema]},
        doc: """
        Either a string or a keyword list with keys:

        #{NimbleOptions.docs(nested_schema, nest_level: 1)}
        """
      ]
    ]

The options for the nested schema in ":my_key" will be nested by
one level, producing the correct Markdown.