Fix formatting of lists
Lists were not correctly indented before,
which means that if you wanted to add
multiline content inside the docs, it would
not be formatted properly and the default
would also be out of place. For example:

    doc: """
    Options are:

      * :foo
      * :bar
    """,
    default: :foo

Would either not indent or add the default
line wrongly:

    Options are:

      * :foo
      * :bar The default value is :bar.

This PR fixes it so lists are naturally indented
and the `nest_level` hack is no longer necessary.
We also remove types printing in complex cases,
because it ends up verbose and/or showing duplicate
information.